Adds UI fixes
Adds dynamic max-height for the todo list container so that when the chat feature appears for Huā, then the container is shorter to leave room for Huā's reply. It then increases height again when the chat is exited. It also adds a fix for the hover-for-chat-button feature. Before, if you hovered over any point width-wise from below the top of Huā, it would trigger the hover as the container for Huā is now 80% of the width for a wider reply from Huā. To solve this, I added a new div for just the chat button and the img for Huā and added the hovering for this container. Slayed.