Updates the background and the border colour for the background
The background had a black line along the side, so this commit crops that out. This isn't an issue with the other graphics, seems to have been a one time export problem. Keep an eye on it in the future though in case it happens again (e.g. with Huā's animations).