Adds and Removes Comments
Adds comments for anyone who might be learning from the code and removes comments which were more personal. Also adds comments which are more personal...but that's because it's relevant for me in the next stages. I need to be able to quickly get a grip of the code again after long times away.